storm/reddit_stream_topology/requirements.txt
cassandra-driver
simplejson